Fix security vulnerabilities
if someone send non-image file (ex: app.py, detect.py, virus something etc), there are two problem
first, if there is file that has same name, it wolud be overwritten so that the project file will be changed.
second, the server will raise error, so that file sent will be remain undeleted, if that file is malware it wolud be serious.